Antigravity/Personal assistant/smartdome-mcp-server/requirements.txt
fastapi
uvicorn
httpx
python-multipart
google-genai
zep-cloud
python-multipart
httpx
